recreation/requirements.txt
appdirs==1.4.3
defusedxml==0.5.0
django-allauth==0.32.0
oauthlib==2.0.2
packaging==16.8
pyparsing==2.2.0
python3-openid==3.1.0
requests-oauthlib==0.8.0
six==1.10.0
